Upload: Change data type of time to float

+++ app/src/main/java/tuni/tuukka/activities/Upload.java
@@ -1,11 +1,9 @@
 package tuni.tuukka.activities;
 
 import android.content.Intent;
-import android.os.AsyncTask;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
-import android.widget.Button;
 import android.widget.TextView;
 
 
@@ -16,7 +14,7 @@ import tuni.tuukka.R;
 public class Upload extends AppCompatActivity {
     private String id;
     private String name;
-    private long time;
+    private float time;
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
@@ -25,7 +23,7 @@ public class Upload extends AppCompatActivity {
         setContentView(R.layout.activity_upload);
         name = extras.getStringExtra("sheetName");
         id = extras.getStringExtra("sheetId");
-        time = extras.getExtras().getLong("time");
+        time = extras.getExtras().getFloat("time");
         ((TextView) findViewById(R.id.upload_sheet_name)).setText(name);
         ((TextView) findViewById(R.id.upload_time_spent)).setText("Time: " + time);
